hw_13/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pydantic[email]
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
fastapi-mail
cloudinary
slowapi
redis
bcrypt==4.0.1
passlib[bcrypt]